refactor:[SMR-2722] calculate gas fee moved (#1735)
