+++ packages/internal/bridge/sdk/src/lib/gas.ts
@@ -1,7 +1,8 @@
 import { FeeData } from '@ethersproject/providers';
 import { BigNumber } from 'ethers';
 
-const doesChainSupportEIP1559 = (feeData: FeeData) => !!feeData.maxFeePerGas && !!feeData.maxPriorityFeePerGas;
+const doesChainSupportEIP1559 = (feeData: FeeData) => !!feeData.maxFeePerGas
+    && !!feeData.maxPriorityFeePerGas && !!feeData.lastBaseFeePerGas;
 
 export const getGasPriceInWei = (feeData: FeeData): BigNumber | null => {
   if (doesChainSupportEIP1559(feeData)) {
@@ -13,3 +14,12 @@ export const getGasPriceInWei = (feeData: FeeData): BigNumber | null => {
   if (feeData.gasPrice) return BigNumber.from(feeData.gasPrice);
   return null;
 };
+
+export function calculateGasFee(
+  feeData: FeeData,
+  gasLimit: number,
+): BigNumber {
+  const gasPriceInWei = getGasPriceInWei(feeData);
+  if (!gasPriceInWei) return BigNumber.from(0);
+  return gasPriceInWei.mul(gasLimit);
+}
